Auth

+++ api/controllers/CartsController.js
@@ -1,9 +1,10 @@
 
 const Cart = require("../models/Cart");
+const Customer = require("../models/User");
 const ObjectId = require('mongoose').Types.ObjectId;
 
 // Forms an array of all the Cart model schema paths excluding only private and protected paths .
-const regex = /(^_)|(^at$)|(^orders$)|(^totalPrice$)/; //Regex that matches private [prefixed with '_'] and protected [those that is not meant to be set by an input .] paths .
+const regex = /(^_)|(^at$)|(^orders$)|(^totalPrice$)|(^customer$)/; //Regex that matches private [prefixed with '_'] and protected [those that is not meant to be set by an input .] paths .
 const schemaPaths = Object.getOwnPropertyNames(Cart.prototype.schema.paths).filter(item => ! regex.test(item));
 
 //Mongoose update options .  
@@ -21,9 +22,20 @@ module.exports.getAll = (req,res,next) => {
 
     ( 
         async  () => {
-         const carts =  await Cart.find().select("-__v").populate('orders',"-__v").lean().exec() ;
-         res.status(200).json(carts);
-         
+
+            let querry = {};
+
+            if( req.user.role !== 'admin' ) //Only admin is authz to get all the carts infos .
+                {
+                     if ( req.user.role !== 'customer' )
+                        throw ( Object.assign(new Error("Forbidden .") , {status : 403}) ); // If not an admin and not a customer you are forbidden to see any carts .
+        
+                    querry = { _id : { $in : req.user.carts } }; //If you the connected user is a customer then it can only observe its own list of carts .
+                }
+
+            const carts =  await Cart.find( querry ).select("-__v").populate('orders customer',"-__v").lean().exec() ;
+            res.status(200).json(carts);
+            
          }
      )
      ().catch(next);
@@ -34,7 +46,8 @@ module.exports.post = (req , res , next) => {
 
     //Declaring newCart , this object will be saved to DB . 
     const newCart = {
-        _id :  new ObjectId() 
+        _id :  new ObjectId() ,
+        customer : req.user._id
     };     
    
     (
@@ -48,9 +61,28 @@ module.exports.post = (req , res , next) => {
             //This helps protect special paths that are not meant to be altered by an input and determined by the backend app logic .
             if( ! require("../functions/isArrEquals")( reqBodyProperties , schemaPaths ) )
                 throw ( Object.assign(new Error("Invalid input .") , {status : 400}) );
- 
-            const cart = await new Cart(newCart).save({select : {__v : -1 }});
+
             
+                //Populating the newCart with values from the request body that matches the schema paths and ignoring other values .
+                schemaPaths.forEach(item => {
+                    if(  req.body[item] != undefined  )
+                        newCart[item] = req.body[item];
+                });
+
+            if( ! ObjectId.isValid(newCart.customer) )
+                throw ( Object.assign(new Error("CUSTOMER ID is invalid .") , {status : 400}) );
+
+            const customer = await Customer.findById(newCart.customer).exec();
+
+            if(customer == null)
+                throw ( Object.assign(new Error("Customer not found .") , {status : 404}) );
+            
+            if( customer.role !== 'customer' )
+               throw ( Object.assign(new Error("Not a customer .") , {status : 400}) );
+
+            const cart = await new Cart(newCart).save({select : {__v : -1 }});
+            await Customer.updateOne({_id : newCart.customer} , {$addToSet : {carts : newCart._id} } , updateOps).exec();//Push the new cart to the list of cart made by the customer . 
+
             res.status(201).json(cart); 
     })()
     .catch(next)
@@ -61,6 +93,7 @@ module.exports.post = (req , res , next) => {
 module.exports.delete = (req , res , next) => {
 
     const cartId = req.params.id ;
+    const customerId = req.user._id ;
 
     (
         async () => {
@@ -69,12 +102,26 @@ module.exports.delete = (req , res , next) => {
             if(! ObjectId.isValid(cartId) )
                 throw ( Object.assign(new Error("Cart ID is invalid .") , {status : 400}) );
 
+            
+            if(! ObjectId.isValid( customerId ) )
+                throw ( Object.assign(new Error("Customer ID is invalid .") , {status : 400}) );    
+
+            const carts = req.user.carts.map(element => 
+                element = element.toString()
+            );
+
+            if( ! carts.includes(cartId) )
+                throw ( Object.assign(new Error("Cart not found .") , {status : 404}) );    
+    
             const deletedCart = await Cart.findByIdAndRemove( cartId , deleteOps ).exec();
 
            if(deletedCart == null)
                 throw ( Object.assign(new Error("Cart not found .") , {status : 404}) );
 
- 
+            
+            await Customer.updateOne({_id : deletedCart.customer} , {$pull : {carts : deletedCart._id}  } , updateOps).exec();//Pull the removed cart from the cart list of the customer . 
+
+
             res.status(201).json(deletedCart);
         }
     )().catch(next)
