+++ api/controllers/ProductsController.js
@@ -4,10 +4,13 @@ const Store = require('../models/Store');
 const ObjectId = require('mongoose').Types.ObjectId;
 const {unlink} = require('fs'); 
 // Forms an array of all the Product model schema paths excluding only private and protected paths .
-const regex = /(^_)|(^imgPath$)|(^createdAt$)|(^orders$)/; //Regex that matches private [prefixed with '_'] and protected [those that is not meant to be set by an input .] paths .
+const regex = /(^_)|(^imgPath$)|(^createdAt$)|(^orders$)|(^store$)/; //Regex that matches private [prefixed with '_'] and protected [those that is not meant to be set by an input .] paths .
 const schemaPaths = Object.getOwnPropertyNames(Product.prototype.schema.paths).filter(item => ! regex.test(item));
 
-//Mongoose update options .  
+//Mongoose update options .
+
+const pageSize = 10 ; // Size of pool products on a page .
+
 const updateOps = {
     useFindAndModify : false ,
     runValidators : true ,
@@ -22,7 +25,13 @@ module.exports.getAll = (req,res,next) => {
 
     ( 
         async  () => {
-         const products =  await Product.find().select("-__v").populate('store orders',"-__v").lean().exec() ;
+
+         const pageNum = Math.min( Math.max( 0 , req.params.page ) , Number.MAX_SAFE_INTEGER );
+
+         if ( isNaN(pageNum) )
+              throw ( Object.assign(new Error("Invalid page number .") , {status : 400}) );
+      
+         const products =  await Product.find().skip( pageSize * pageNum ).limit( pageSize ).select("-__v").populate('store orders',"-__v").lean().exec() ;
          res.status(200).json(products);
          
          }
@@ -34,7 +43,8 @@ module.exports.post = (req , res , next) => {
 
     //Declaring newProduct , this object will be saved to DB . 
     const newProduct = {
-        _id : req.id || new ObjectId() 
+        _id : req.id || new ObjectId() ,
+        store : req.user.store
     };     
    
     (
@@ -61,8 +71,12 @@ module.exports.post = (req , res , next) => {
                  newProduct[item] = req.body[item];
          });
      
-        //Tests weither the given ID in the URL can be a valid ObjectID or not , in case it cannot be a valid ObjectID an error with status 400 is returned and no need to query the DB .
-        if(! ObjectId.isValid(newProduct.store) )
+         //Tests weither the given ID in the URL can be a valid ObjectID or not , in case it cannot be a valid ObjectID an error with status 400 is returned and no need to query the DB .
+      
+         if( newProduct.store == undefined )
+          throw ( Object.assign(new Error("You do not own a store .") , {status : 400}) );
+      
+         if(! ObjectId.isValid(newProduct.store) )
             throw ( Object.assign(new Error("Store ID is invalid .") , {status : 400}) );
 
         const store = await Store.findById(newProduct.store).exec();
@@ -95,6 +109,7 @@ module.exports.post = (req , res , next) => {
 module.exports.put = (req , res , next) => {
 
     const productId = req.params.id ;
+    const storeId = req.user.store ;
     const updateProduct = {} ;
 
 
@@ -105,11 +120,32 @@ module.exports.put = (req , res , next) => {
             //If an image is uploaded then its path must be included in the newProduct POJO to be saved in the DB .
             if(req.file != undefined )
                updateProduct.imgPath = req.file.path.replace(/\\/g,"/"); 
-                 
+             
+            if( storeId == undefined )
+               throw ( Object.assign(new Error("You do not own a store .") , {status : 400}) );
+           
+            if(! ObjectId.isValid(storeId) )
+                 throw ( Object.assign(new Error("Store ID is invalid .") , {status : 400}) );
 
             //Tests weither the given ID in the URL can be a valid ObjectID or not , in case it cannot be a valid ObjectID an error with status 400 is returned and no need to query the DB .
             if(! ObjectId.isValid(productId) )
-            throw ( Object.assign(new Error("Product ID is invalid .") , {status : 400}) );
+               throw ( Object.assign(new Error("Product ID is invalid .") , {status : 400}) );
+
+            // The store and Product ID are valid , this block will querry the DB to find the store doc and checks if the productID is an ID of a product owned by the connected sotre owner store .
+            const result = await Store.findById( storeId ).lean().exec() ;
+                          
+            if(result == null)
+                  throw ( Object.assign(new Error(` Store not found .`) , {status : 404}) );//If no document is found a not found response is sent back with 404 status code .
+            
+            // products property of store doc is ann array of ObjectId , this block casts each ObjectId to its string value to check if the string value of the ProductId is invcluded in the products .
+            result.products = result.products.map(element => 
+                    element = element.toString()
+                );
+
+            if (! result.products.includes(productId) )
+              throw ( Object.assign(new Error(` Not found .`) , {status : 404}) );//If no document is found a not found response is sent back with 404 status code .
+        
+
 
             const reqBodyProperties = Object.getOwnPropertyNames(req.body);//populate reqBodyProperties with req.body property names .
             //Tests weither the req.body contains properties that respects the schema , in case there is at least one invalid property name an error of status 400 will be returned .
@@ -139,13 +175,35 @@ module.exports.put = (req , res , next) => {
 module.exports.delete = (req , res , next) => {
 
     const productId = req.params.id ;
+    const storeId = req.user.store ;
 
     (
         async () => {
             
+            if( storeId == undefined )
+               throw ( Object.assign(new Error("You do not own a store .") , {status : 400}) );
+           
+            if(! ObjectId.isValid(storeId) )
+                 throw ( Object.assign(new Error("Store ID is invalid .") , {status : 400}) );
+
+
             //Tests weither the given ID in the URL can be a valid ObjectID or not , in case it cannot be a valid ObjectID an error with status 400 is returned and no need to query the DB .
             if(! ObjectId.isValid(productId) )
-            throw ( Object.assign(new Error("Product ID is invalid .") , {status : 400}) );
+                throw ( Object.assign(new Error("Product ID is invalid .") , {status : 400}) );
+
+            const result = await Store.findById( storeId ).lean().exec() ;
+                          
+            if(result == null)
+                  throw ( Object.assign(new Error(` Store not found .`) , {status : 404}) );//If no document is found a not found response is sent back with 404 status code .
+            
+            result.products = result.products.map(element => 
+                    element = element.toString()
+                );
+
+            if (! result.products.includes(productId) )
+              throw ( Object.assign(new Error(` Not found .`) , {status : 404}) );//If no document is found a not found response is sent back with 404 status code .
+        
+
 
             const deletedProduct = await Product.findByIdAndRemove( productId , deleteOps ).exec();
